Adicionado botão relacionar e falhas de lógica corrigidas

+++ formulario.ps1
@@ -1,6 +1,7 @@
 #Assemblies
 Add-Type -AssemblyName System.Windows.forms
 Add-Type -Assemblyname System.Drawing
+Add-Type -AssemblyName PresentationFramework
 
 #menu
 #Janela principal menu
@@ -38,8 +39,14 @@ $inserirLoja.Location = New-Object System.Drawing.Size(20,90)
 $inserirLoja.Size = New-Object System.Drawing.Size(130,20)
 $inserirLoja.Text = "Criar loja"
 $inserirLoja.Add_click({
-    $form1.ShowDialog()
-   
+    $caixa1.Text=""
+    $result=$formLoja.ShowDialog()
+    if($result -eq "Cancel"){
+        popularLojas
+    }
+    $msgBottom.Text= ""
+    $dMsgBottom.Text= ""
+    
 })
 $menu.Controls.Add($inserirLoja)
 
@@ -49,7 +56,13 @@ $inserirDepartamento.Location = New-Object System.Drawing.Size(180,90)
 $inserirDepartamento.Size = New-Object System.Drawing.Size(130,20)
 $inserirDepartamento.Text = "Criar departamento"
 $inserirDepartamento.Add_click({
-    $form2.ShowDialog()  
+    $dcaixa1.Text=""
+    $formDepartamento.ShowDialog() 
+    popularLojas
+    $listBoxSaida.items.Clear() 
+    $msgBottom.Text= ""
+    $dMsgBottom.Text= ""
+    
 })
 $menu.Controls.Add($inserirDepartamento)
 
@@ -67,34 +80,15 @@ $LabelDepartamentos.Location =  New-Object System.Drawing.Size(9,330)
 $LabelDepartamentos.AutoSize = $true
 $menu.Controls.Add($LabelDepartamentos)
 
-#botão Associar departamentos
-$botaoAssociarDepartamento = New-Object System.Windows.Forms.Button 
-$botaoAssociarDepartamento.Location = New-Object System.Drawing.Size(290,200)
-$botaoAssociarDepartamento.Size = New-Object System.Drawing.Size(130,20)
-$botaoAssociarDepartamento.Text = "Associar departamento"
-$botaoAssociarDepartamento.Add_click({
+#botão Teste
+$botaoTeste = New-Object System.Windows.Forms.Button 
+$botaoTeste.Location = New-Object System.Drawing.Size(290,200)
+$botaoTeste.Size = New-Object System.Drawing.Size(130,20)
+$botaoTeste.Text = "Botão teste"
+$botaoTeste.Add_click({
 
-    #$novaLinha= $listBoxLojas.SelectedItem -split "\|")[0] + $listBoxSaida.SelectedItem -split "\|")[0]
-
-
-
-
-    
-    
-
-
-
-
-    #if ($listBoxLojas.SelectedIndex -ne -1){
-    #    $selecao=$listBoxLojas.SelectedIndex + 1
-    #    $mLabel.Text = "Selecionado $selecao"
-    #}
-    #else {
-    #    $mLabel.Text = "Não selecionado"
-    #}
-    
 })
-$menu.Controls.Add($botaoAssociarDepartamento)
+#$menu.Controls.Add($botaoTeste)
 
 #botão relacionar
 $botaoRelacionar = New-Object System.Windows.Forms.Button 
@@ -103,73 +97,85 @@ $botaoRelacionar.Size = New-Object System.Drawing.Size(130,20)
 $botaoRelacionar.Text = "Criar relacionamento"
 $botaoRelacionar.Add_click({
 
+    #[System.Windows.MessageBox]::Show('Opção em desenvolvimento')
 
+    if ($null -eq $listBoxLojas.SelectedItem -or $null -eq $listBoxSaida.SelectedItem){
+        [System.Windows.MessageBox]::Show('Selecione uma loja e um departamento.', 'Erro')
+    }
+    else{
+        $lojaAdicionada=($listBoxLojas.SelectedItem -split "\|")[0]
+        $departamentoAdicionado=($listBoxSaida.SelectedItem -split "\|")[0]
+        Add-Content -Value "$lojaAdicionada|$departamentoAdicionado" -Path .\relacionamentos.txt
+        exibeDepartamentos
+        [System.Windows.MessageBox]::Show('Relacionamento criado')
+
+    }
 
 })
 $menu.Controls.Add($botaoRelacionar)
 
 #Janela Inserção de loja
-$form1 = New-Object System.Windows.Forms.Form
-$form1.text = "Inserção de loja"
-$Form1.size = New-Object System.Drawing.Size(350,200)
-$Form1.StartPosition = "CenterScreen"
+$formLoja = New-Object System.Windows.Forms.Form
+$formLoja.text = "Inserção de loja"
+$formLoja.size = New-Object System.Drawing.Size(350,200)
+$formLoja.StartPosition = "CenterScreen"
 
 #Janela Inserção de departamento
-$form2 = New-Object System.Windows.Forms.Form
-$form2.text = "Inserção de departamento"
-$Form2.size = New-Object System.Drawing.Size(350,200)
-$Form2.StartPosition = "CenterScreen"
+$formDepartamento = New-Object System.Windows.Forms.Form
+$formDepartamento.text = "Inserção de departamento"
+$formDepartamento.size = New-Object System.Drawing.Size(350,200)
+$formDepartamento.StartPosition = "CenterScreen"
 
 
 #Caixa de texto
 $caixa1 = New-Object System.Windows.Forms.TextBox
 $caixa1.Location = New-Object System.Drawing.Size(100,50)
 $caixa1.Size = New-Object System.Drawing.Size(208,20)
-$form1.Controls.Add($caixa1) #adiciona a caixa de texto na janela grafica
+$formLoja.Controls.Add($caixa1) #adiciona a caixa de texto na janela grafica
 #Caixa de texto
-$dcaixa1 = New-Object System.Windows.Forms.TextBox
-$dcaixa1.Location = New-Object System.Drawing.Size(100,50)
-$dcaixa1.Size = New-Object System.Drawing.Size(208,20)
-$form2.Controls.Add($dcaixa1) #adiciona a caixa de texto na janela grafica
+$dCaixa1 = New-Object System.Windows.Forms.TextBox
+$dCaixa1.Location = New-Object System.Drawing.Size(100,50)
+$dCaixa1.Size = New-Object System.Drawing.Size(208,20)
+$formDepartamento.Controls.Add($dCaixa1) #adiciona a caixa de texto na janela grafica
 
 #Label loja
 $label = New-Object System.Windows.Forms.label
 $label.Text = "Digite o nome da loja a ser criada:"
 $label.Location =  New-Object System.Drawing.Size(20,15)
 $label.AutoSize = $true
-$form1.Controls.Add($label)
+$formLoja.Controls.Add($label)
 
 #Label departamento
 $dlabel = New-Object System.Windows.Forms.label
 $dlabel.Text = "Digite o nome do departamento a ser criado:"
 $dlabel.Location =  New-Object System.Drawing.Size(20,15)
 $dlabel.AutoSize = $true
-$form2.Controls.Add($dlabel)
+$formDepartamento.Controls.Add($dlabel)
 
 
 
 #Mensagem parte inferior loja
-$MsgBottom = New-Object System.Windows.Forms.label
-$MsgBottom.Location =  New-Object System.Drawing.Size(20,130)
-$MsgBottom.AutoSize = $true
-$form1.Controls.Add($MsgBottom)
+$msgBottom = New-Object System.Windows.Forms.label
+$msgBottom.Location =  New-Object System.Drawing.Size(20,130)
+$msgBottom.AutoSize = $true
+$formLoja.Controls.Add($msgBottom)
 
 #Mensagem parte inferior departamento
 $dMsgBottom = New-Object System.Windows.Forms.label
 $dMsgBottom.Location =  New-Object System.Drawing.Size(20,130)
 $dMsgBottom.AutoSize = $true
-$form2.Controls.Add($dMsgBottom)
+$formDepartamento.Controls.Add($dMsgBottom)
 
 #idLoja label
 $idLoja = New-Object System.Windows.Forms.label
 $idLoja.Location =  New-Object System.Drawing.Size(23,52)
 $idLoja.AutoSize = $true
-$form1.Controls.Add($idLoja)
+$formLoja.Controls.Add($idLoja)
 #idDepartamento label
 $idDepartamento = New-Object System.Windows.Forms.label
 $idDepartamento.Location =  New-Object System.Drawing.Size(23,52)
 $idDepartamento.AutoSize = $true
-$form2.Controls.Add($idDepartamento)
+$formDepartamento.Controls.Add($idDepartamento)
 
 
 #botão cancelar loja
@@ -177,16 +183,16 @@ $canc = New-Object System.Windows.Forms.Button
 $canc.Location = New-Object System.Drawing.Size(130,90)
 $canc.Size = New-Object System.Drawing.Size(100,20)
 $canc.Text = "Fechar"
-$canc.Add_Click({$form1.Tag = $form1.close()}) 
-$form1.Controls.Add($canc)
+$canc.Add_Click({$formLoja.Tag = $formLoja.close()}) 
+$formLoja.Controls.Add($canc)
 
 #botão cancelar departamento
 $dcanc = New-Object System.Windows.Forms.Button 
 $dcanc.Location = New-Object System.Drawing.Size(130,90)
 $dcanc.Size = New-Object System.Drawing.Size(100,20)
 $dcanc.Text = "Fechar"
-$dcanc.Add_Click({$form2.Tag = $form2.close()}) 
-$form2.Controls.Add($dcanc)
+$dcanc.Add_Click({$formDepartamento.Tag = $formDepartamento.close()}) 
+$formDepartamento.Controls.Add($dcanc)
 
 #botão ok loja
 $ok = New-Object System.Windows.Forms.Button 
@@ -195,25 +201,39 @@ $ok.Size = New-Object System.Drawing.Size(100,20)
 $ok.Text = "Ok"
 $ok.Add_click({ #essa parte é executada ao clicar no botão ok
     if($caixa1.text -eq ""){
-        $MsgBottom.Text = "Loja não adicionada, caixa de texto vazia"
+        $msgBottom.Text = "Loja não adicionada, caixa de texto vazia"
     }
     else{
-        $ultimo=Get-Content .\ixTbLoja.txt #variável recebe conteúdo do texto
-        $texto=$caixa1.text
-        Add-Content -Value "$ultimo|$texto" -Path .\tbLoja.txt #nova linha adicionada no arquivo
+        $cont=0
+        foreach ($linha in Get-Content .\tbLoja.txt){ #verifica se já existe algum com o mesmo nome
+            if(($linha -split "\|")[1] -eq $caixa1.text){
+                $cont++
+                break
+            }
 
-        [int]$ultimo=$ultimo #variável é convertida para int
-        $ultimo++ #e é incrementada
-        [string]$ultimo=([string]$ultimo).PadLeft(4,'0') #variável volta a ser string padronizada com zeros à esquerda
-        Clear-Content -Path .\ixTbLoja.txt 
-        Add-Content -Value $ultimo -Path .\ixTbLoja.txt 
-        $idLoja.Text = "codLoja: " + $ultimo + ":"
-        $MsgBottom.Text = "Loja $texto inserida."
-        $caixa1.text= ""
+        }
+        if($cont -eq 0){
+            $ultimo=Get-Content .\ixTbLoja.txt #variável recebe conteúdo do texto
+            $texto=$caixa1.text
+            Add-Content -Value "$ultimo|$texto" -Path .\tbLoja.txt #nova linha adicionada no arquivo
+
+            [int]$ultimo=$ultimo #variável é convertida para int
+            $ultimo++ #e é incrementada
+            [string]$ultimo=([string]$ultimo).PadLeft(4,'0') #variável volta a ser string padronizada com zeros à esquerda
+            Clear-Content -Path .\ixTbLoja.txt 
+            Add-Content -Value $ultimo -Path .\ixTbLoja.txt 
+            $idLoja.Text = "codLoja: " + $ultimo + ":"
+            $msgBottom.Text = "Loja $texto inserida."
+            $caixa1.text= ""
+        }
+        else{
+            $msgBottom.Text = "Loja já existente"
+        }
+        
     }
 
 })
-$form1.Controls.Add($ok)
+$formLoja.Controls.Add($ok)
 
 #botão ok departamento
 $dok = New-Object System.Windows.Forms.Button #botão ok
@@ -221,45 +241,48 @@ $dok.Location = New-Object System.Drawing.Size(20,90)
 $dok.Size = New-Object System.Drawing.Size(100,20)
 $dok.Text = "Ok"
 $dok.Add_click({ #essa parte é executada ao clicar no botão ok
-    if($dcaixa1.text -eq ""){
+    if($dCaixa1.text -eq ""){
         $dMsgBottom.Text = "Departamento não adicionado, caixa de texto vazia"
     }
     else{
-        $dultimo=Get-Content .\ixTbDepartamento.txt #variável recebe conteúdo do texto
-        $dtexto=$dcaixa1.text
-        Add-Content -Value "$dultimo|$dtexto" -Path .\tbDepartamento.txt #nova linha adicionada no arquivo
+        $cont=0
+        foreach ($linha in Get-Content .\tbDepartamento.txt){ #verifica se já existe algum com o mesmo nome
+            if(($linha -split "\|")[1] -eq $dCaixa1.text){
+                $cont++
+                break
+            }
 
-        [int]$dultimo=$dultimo #variável é convertida para int
-        $dultimo++ #e é incrementada
-        [string]$dultimo=([string]$dultimo).PadLeft(4,'0') #variável volta a ser string padronizada com zeros à esquerda
-        Clear-Content -Path .\ixTbDepartamento.txt 
-        Add-Content -Value $dultimo -Path .\ixTbDepartamento.txt 
-        $idDepartamento.Text = "codDpt: " + $dultimo + ":"
-        $dMsgBottom.Text = "Departamento $dtexto inserido."
-        $dcaixa1.text= ""
+        }
+        if($cont -eq 0){
+            $dultimo=Get-Content .\ixTbDepartamento.txt #variável recebe conteúdo do texto
+            $dtexto=$dCaixa1.text
+            Add-Content -Value "$dultimo|$dtexto" -Path .\tbDepartamento.txt #nova linha adicionada no arquivo
+
+            [int]$dultimo=$dultimo #variável é convertida para int
+            $dultimo++ #e é incrementada
+            [string]$dultimo=([string]$dultimo).PadLeft(4,'0') #variável volta a ser string padronizada com zeros à esquerda
+            Clear-Content -Path .\ixTbDepartamento.txt 
+            Add-Content -Value $dultimo -Path .\ixTbDepartamento.txt 
+            $idDepartamento.Text = "codDpt: " + $dultimo + ":"
+            $dMsgBottom.Text = "Departamento $dtexto inserido."
+            $dCaixa1.text= ""
+        }
+        else{
+            $dMsgBottom.Text = "Departamento já existente"
+        }
+
+
+        
     }
 })
-$form2.Controls.Add($dok)
+$formDepartamento.Controls.Add($dok)
 
+function popularLojas{ #preenche a listbox de lojas
+    $listBoxLojas.items.Clear()
+    foreach ($linha in Get-Content .\tbLoja.txt){ [void]$listBoxLojas.Items.Add($linha) } 
+}
 
-
-
-
-#
-#
-#Início do programa
-#
-#
-$idLoja.Text = "codLoja: " + (Get-Content .\ixTbLoja.txt) + ":" #preenche o label
-$idDepartamento.Text = "codDpt: " + (Get-Content .\ixTbDepartamento.txt) + ":" #preenche o label
-foreach ($linha in Get-Content .\tbLoja.txt){ [void]$listBoxLojas.Items.Add($linha) } #preenche a listbox
-
-
-
-$listBoxLojas.add_SelectedIndexChanged({ #Essa parte é executada quando o usuário clica em uma loja
-    #$sel=$listBoxLojas.SelectedIndex
-    #$mLabel.Text = "index $sel"
-
+function exibeDepartamentos{
     $listBoxSaida.items.Clear()
 
 
@@ -274,14 +297,36 @@ $listBoxLojas.add_SelectedIndexChanged({ #Essa parte é executada quando o usuári
         }
     }
 
+}
+
+
+
+
+#
+#
+#Início do programa
+#
+#
+$idLoja.Text = "codLoja: " + (Get-Content .\ixTbLoja.txt) + ":" #preenche o label
+$idDepartamento.Text = "codDpt: " + (Get-Content .\ixTbDepartamento.txt) + ":" #preenche o label
+popularLojas
+
+
+
+
+$listBoxLojas.add_SelectedIndexChanged({ #Essa parte é executada quando o usuário clica em uma loja
+    #$sel=$listBoxLojas.SelectedIndex
+    #$mLabel.Text = "index $sel"
+
+    exibeDepartamentos
+
 
     #$listBoxSaida.items.add(($listBoxLojas.text -split "\|")[1])
     
 })
 
-if($null -eq $listboxsaida.SelectedItem){
-    $mLabel.Text = "sim"
-}
+
+
 
 [void]$menu.ShowDialog()
 
@@ -289,8 +334,8 @@ if($null -eq $listboxsaida.SelectedItem){
 
 
 #$mLabel.Text =$listBoxLojas.SelectedIndex
-#$form1.ShowDialog()#mostra a janela gráfica na tela
-#$form2.ShowDialog()
+#$formLoja.ShowDialog()#mostra a janela gráfica na tela
+#$formDepartamento.ShowDialog()
 #$idLoja.Text ="codLoja: " + ((Get-Content .\tbLoja.txt -Tail 1) -split "\|")[0]
